Support for printing information for both left and right subsystem.
If the total state is thought of as AxB where A is the left subsystem
and B is the right subsystem, then information can now also be printed
for B. An application for this would be if you have some qubits of
interest and som ancillary qubits:

register q0[0]
register q1[1]
register a0[0]
register a1[1]

Then you can print either the states and probability distribution for
the sybsystem A=q0xq1 or B=a0xa1.

Break out code to print mixture into its own function for clarity.

Document  parameter.